requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
nltk>=3.8
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0